Multi-option repository page
Breaks the form on this page into two sections, one for copying the
selected datasets to another endpoint and the other for actually
graphing them.

Because we need the checked/unchecked state of the individual datasets
in the table, we must continue to use a single `<form>` that wraps both
the table and the actual form controls below.