Use client secrets JSON file for Auth config
This enables the use of the simpler `flow_from_clientsecrets()` function
for creating a Google OAuth2 "flow" object based on the configuration
stored in `auth.json`.

Because we still have some custom calls (notably logout) that also needs
access to the same Auth configuration information, it is re-exported via
`GLOBUS_AUTH` in the Flask `app.config` lookup.